+++ .gitignore
@@ -5,12 +5,11 @@
 #*.pdf
 #*.zip
 /**/venv/*
-
 /**/__pycache__/*
-**/interpolation_integration_generator/static/interpolation_integration_generator/*/*.tex
-**/interpolation_integration_generator/static/interpolation_integration_generator/*/*.pdf
-**/interpolation_integration_generator/static/interpolation_integration_generator/*/*.log
-**/interpolation_integration_generator/static/interpolation_integration_generator/*/*.aux
-**/interpolation_integration_generator/static/interpolation_integration_generator/*/*.fls
-**/interpolation_integration_generator/static/interpolation_integration_generator/*/*.fdb_latexmk
-**/interpolation_integration_generator/static/interpolation_integration_generator/*/*.zip
+**/generator/static/generator/variants/*/*.tex
+**/generator/static/generator/variants/*/*.pdf
+**/generator/static/generator/variants/*/*.log
+**/generator/static/generator/variants/*/*.aux
+**/generator/static/generator/variants/*/*.fls
+**/generator/static/generator/variants/*/*.fdb_latexmk
+**/generator/static/generator/variants/*/*.zip
