+++ Generator/Generator/urls.py
@@ -1,0 +1,5 @@
+from django.urls import path, include
+
+urlpatterns = [
+    path('', include('generator.urls')),
+]
