+++ Generator/Tasks/Polynomial.py
