+++ Generator/Tasks/PolynomialHelper.py
