+++ Generator/generator/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class InterpolationIntegrationGeneratorConfig(AppConfig):
-    name = 'interpolation_integration_generator'
+    name = 'generator'
