+++ Generator/generator/static/generator/css/result_page.css
