+++ Generator/generator/static/generator/js/control_checkboxes.js
