+++ Generator/generator/static/generator/js/relation_between_file_field_and_variants.js
