+++ Generator/generator/static/generator/js/relations_between_variants_and_max_degree.js
