+++ Generator/generator/templates/generator/index.html
@@ -6,7 +6,7 @@
     <meta charset="UTF-8">
     <title>Generator</title>
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
-    <link rel="stylesheet" href="{% static 'interpolation_integration_generator/css/result_page.css' %}">
+    <link rel="stylesheet" href="{% static 'generator/css/result_page.css' %}">
 </head>
 <body>
     <div class="btn-toolbar">
