+++ Generator/generator/templates/generator/result_page.html
@@ -5,7 +5,7 @@
 <head>
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
           integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <link rel="stylesheet" href="{% static 'interpolation_integration_generator/css/result_page.css' %}">
+    <link rel="stylesheet" href="{% static 'generator/css/result_page.css' %}">
     <meta charset="UTF-8">
     <title>Result page</title>
 </head>
