+++ Generator/generator/templates/generator/splines.html
@@ -10,7 +10,7 @@
 <head>
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
           integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <link rel="stylesheet" href="{% static 'interpolation_integration_generator/css/result_page.css' %}">
+    <link rel="stylesheet" href="{% static 'generator/css/result_page.css' %}">
     <meta charset="UTF-8">
     <title>Result page</title>
 </head>
@@ -19,10 +19,10 @@
 <div class="container">
     <div class="row">
         <div class="col mr-4">
-            <form action="{% url 'generate_interpolation' %}" method="post" enctype="multipart/form-data"
-                  class="col mr-4">
+            <form action="{% url 'generate_splines' %}" method="post" enctype="multipart/form-data" class="col mr-4">
                 {% csrf_token %}
-                <h3 class="mt-2">Интерполяционный многочлен Ньютона и Лагранжа</h3>
+                <h3 class="mt-2">Генератор заданий по сплайнам</h3>
+                <br>
                 <div class="row mt-2">
                     <div class="col">
                         <h5>{{ form.filename.label }}</h5>
@@ -57,36 +57,6 @@
                     </div>
                 </div>
                 <br>
-                <div class="row mt-2">
-                    <div class="col">
-                        <h5>{{ form.number_of_variants_in_string.label }}</h5>
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="col">
-                        {% for choice in form.number_of_variants_in_string %}
-                            <div>
-                                {{ choice }}
-                            </div>
-                        {% endfor %}
-                    </div>
-                </div>
-                <br>
-                <div class="row mt-2">
-                    <div class="col">
-                        <h5>{{ form.the_biggest_polynomial_degree.label }}</h5>
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="col">
-                        {% for choice in form.the_biggest_polynomial_degree %}
-                            <div>
-                                {{ choice }}
-                            </div>
-                        {% endfor %}
-                    </div>
-                </div>
-                <br>
 
                 <div class="row mt-2">
                     <div class="col">
@@ -103,6 +73,60 @@
                     </div>
                 </div>
                 <br>
+
+                <div class="row">
+                    <div class="col">
+                        <h5>{{ form.x1.label }}</h5>
+                    </div>
+                    <div class="col">
+                        {{ form.x1 }}
+                    </div>
+                </div>
+
+                <div class="row">
+                    <div class="col">
+                        <h5>{{ form.x2.label }}</h5>
+                    </div>
+                    <div class="col">
+                        {{ form.x2 }}
+                    </div>
+                </div>
+
+                <div class="row">
+                    <div class="col">
+                        <h5>{{ form.y1.label }}</h5>
+                    </div>
+                    <div class="col">
+                        {{ form.y1 }}
+                    </div>
+                </div>
+
+                <div class="row">
+                    <div class="col">
+                        <h5>{{ form.y2.label }}</h5>
+                    </div>
+                    <div class="col">
+                        {{ form.y2 }}
+                    </div>
+                </div>
+                <br>
+
+                 <div class="row mt-2">
+                    <div class="col">
+                        <h5>{{ form.step.label }}</h5>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col">
+                        {% for choice in form.step %}
+                            <div>
+                                {{ choice }}
+                            </div>
+                        {% endfor %}
+                    </div>
+                </div>
+                <br>
+
                 <em>Необязательный параметр</em>
                 <div class="row">
                     <div class="col">
@@ -112,15 +136,14 @@
                         {{ form.seed }}
                     </div>
                 </div>
-
-                <button type="submit" id="submit" class="btn mb-2 mt-2 btn-primary btn-lg active" role="button" aria-pressed="true">
-                    Сгенерировать
+                <button type="submit" class="btn mb-2 btn-primary btn-lg active" role="button" aria-pressed="true"
+                        id="submit">Сгенерировать
                 </button>
             </form>
+
         </div>
     </div>
 </div>
 </body>
-<script src="{% static 'interpolation_integration_generator/js/relations_between_variants_and_max_degree.js' %}"></script>
 <script src="{% static 'interpolation_integration_generator/js/relation_between_file_field_and_variants.js' %}"></script>
 </html>
