--- InterpolationIntegrationGenerator/Generator/urls.py
@@ -1,5 +1,0 @@
-from django.urls import path, include
-
-urlpatterns = [
-    path('', include('interpolation_integration_generator.urls')),
-]
